requirements.txt
altair==4.2.0
annexremote==1.6.0
appdirs==1.4.4
APScheduler==3.10.0
astroid==2.13.2
attrs==22.2.0
black==23.1.0
blinker==1.5
bokeh==2.4.3
boto==2.49.0
cachetools==5.2.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
datalad==0.18.0
datalad-metadata-model==0.3.7
datalad-metalad==0.4.10
decorator==5.1.1
dill==0.3.6
distro==1.8.0
dnspython==2.3.0
docopt==0.6.2
entrypoints==0.4
fasteners==0.18
flake8==6.0.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
graphviz==0.20.1
greenlet==2.0.1
humanize==4.4.0
idna==3.4
importlib-metadata==6.0.0
iso8601==1.1.0
isort==5.11.4
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
keyrings.alt==4.2.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
looseversion==1.0.3
MarkupSafe==2.1.1
matplotlib==3.6.3
mccabe==0.7.0
more-itertools==9.0.0
msgpack==1.0.4
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.1
obfuskey==0.1.3
packaging==23.0
pandas==1.5.2
pathspec==0.11.0
patool==1.12
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.6.2
protobuf==3.20.3
pyarrow==10.0.1
pycallgraph2==1.1.3
pycodestyle==2.10.0
pydeck==0.8.0
pydeps==1.11.1
pyflakes==3.0.1
Pygments==2.14.0
# pygraphviz==1.10
pylint==2.15.10
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-gitlab==3.12.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.0.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.46
stdlib-list==0.8.0
streamlit==1.17.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
toposort==1.9
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
wrapt==1.14.1
yarg==0.1.9
zipp==3.11.0
